requirements.txt
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.1
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
feedparser==6.0.12
Flask==3.0.0
greenlet==3.3.0
groq==1.0.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
loguru==0.7.3
MarkupSafe==3.0.3
numpy==2.3.5
packaging==25.0
praw==7.8.1
prawcore==2.4.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
update-checker==0.18.0
urllib3==2.6.2
websocket-client==1.9.0
Werkzeug==3.1.4
win32_setctime==1.2.0
